Change an int to a size_t in save.cpp to allow using more than 2^31 particles.

+++ save.cpp
@@ -117,7 +117,7 @@ template <typename T> class BufferedWrite
             return written;
        }
     protected:
-       virtual T setter(int i, int k, int type) = 0;
+       virtual T setter(size_t i, int k, int type) = 0;
        virtual ~BufferedWrite()
        {
            free(block);
@@ -164,7 +164,7 @@ class PosBufferedWrite : public BufferedWrite<float>
                     return "POS ";
             }
         }
-        virtual float setter(int i, int k, int type)
+        virtual float setter(size_t i, int k, int type)
         {
           double value = Pgrid.Pos(i,k, type) + Pgrid.get_shift(type);
           if(outdata)
@@ -198,7 +198,7 @@ class VelBufferedWrite : public BufferedWrite<float>
                     return "VEL ";
             }
         }
-        virtual float setter(int i, int k, int type)
+        virtual float setter(size_t i, int k, int type)
         {
           if(k == 0)
               get_new_therm_vels();
@@ -240,7 +240,7 @@ class IDBufferedWrite : public BufferedWrite<id_type>
                     return "ID  ";
             }
         }
-        virtual id_type setter(int i, int k, int type)
+        virtual id_type setter(size_t i, int k, int type)
         {
             return i + FirstId;
         }
@@ -255,7 +255,7 @@ class EnergyBufferedWrite : public BufferedWrite<float>
         BufferedWrite(snap, NumPart, 1, snap.GetFormat() > 2 ? "InternalEnergy" : "U   ")
         {}
     private:
-    virtual float setter(int i, int k, int type)
+    virtual float setter(size_t i, int k, int type)
     {
         return 0;
     }
